[ft] Major overhaul to how threads work.
The core scheduler no longer knows anything about threads, it merely
schedules jobs. A job is anything with a call operator and context.

This is in preparation to support for C++20 coroutines. It also makes
the core more portable since now we can at least build to platforms
without setjmp/longjmp (WASM for instance).