[ft] async semaphore wait manages interrupts
To perform the counter modify, we need to disable interrupts since the
semaphore is possibly shared with an interrupt handler.

However, normally the thread blocks and calls suspend_self, which
eventually re-enables interrupts for the next job.

With a coroutine, the control just goes back to a caller, and interrupts
remain disabled. This commit manually re-enables interrupts before
returning and re-disables them before resumption to construct the
intended behavior.